LayoutTests/svg/custom/stroke-width-click-expected.txt


  
  
    if (window.layoutTestController)
      layoutTestController.dumpAsText();
    
    function log(string) {
      var newDiv = document.createElementNS("http://www.w3.org/1999/xhtml", "div");
      newDiv.textContent = string;
      document.getElementById('log').appendChild(newDiv);
    }
    
    var lastEvent = null;
    
    function testClick(x, y, expected) {
      eventSender.mouseMoveTo(x, y);
      eventSender.mouseDown();
      eventSender.mouseUp();
      if (expected && lastEvent && lastEvent.clientX == x && lastEvent.clientY == y) {
        log("SUCCESS: click received at: " + x + ", " + y);
      } else if (!expected && !lastEvent) {
        log("SUCCESS: click at " + x + ", " + y + " was correctly ignored");
      } else if (expected && lastEvent) {
        log("FAILURE: click received at: " + lastEvent.clientX + ", " + lastEvent.clientY + " but it was expected at: " + x + ", " + y);
      } else if (!expected && lastEvent) {
        log("FAILURE: UNEXPECTED click received at: " + lastEvent.clientX + ", " + lastEvent.clientY);
      } else if (expected && !lastEvent) {
        log("FAILURE: NO click received at: " + x + ", " + y);
      } else {
        log("LOGIC ERORR");
      }
      lastEvent = null;
    }
    
    function gotClick(evt) {
      lastEvent = evt;
    }
    
    function gotMouseEvent(evt, type) {
      if (!window.layoutTestController)
        document.getElementById('mouseover-text').textContent = type + " at " + evt.clientX + ", " + evt.clientY;
    }
  
    
  
  
    
      
      SUCCESS: click at 10, 10 was correctly ignoredSUCCESS: click received at: 30, 100SUCCESS: click at 29, 100 was correctly ignoredSUCCESS: click received at: 30, 50SUCCESS: click at 30, 49 was correctly ignoredSUCCESS: click received at: 30, 150SUCCESS: click at 30, 151 was correctly ignoredSUCCESS: click received at: 300, 100SUCCESS: click at 301, 100 was correctly ignoredSUCCESS: click received at: 300, 50SUCCESS: click at 300, 49 was correctly ignoredSUCCESS: click received at: 300, 150SUCCESS: click at 300, 151 was correctly ignoredSUCCESS: click at 1000, 1000 was correctly ignored
    
  
  
  
  
  
  
  
  
  
    if (window.eventSender) {
      testClick(10, 10, false);
      testClick(30, 100, true);
      testClick(29, 100, false);
      testClick(30, 50, true);
      testClick(30, 49, false);
      testClick(30, 150, true);
      testClick(30, 151, false);
      testClick(300, 100, true);
      testClick(301, 100, false);
      testClick(300, 50, true);
      testClick(300, 49, false);
      testClick(300, 150, true);
      testClick(300, 151, false);
      testClick(1000, 1000, false);
    } else {
      log("ERROR: window.eventSender not found!  This test must be run using DumpRenderTree.  Mousing over the yellow area will log however.");
    }
  
  
